+++ flatlaf-demo/src/main/java/com/formdev/flatlaf/demo/FlatLafDemo.java
@@ -46,11 +46,7 @@ public class FlatLafDemo
 			ex.printStackTrace();
 
 			// fallback
-			try {
-				UIManager.setLookAndFeel( new FlatLightLaf() );
-			} catch( Exception ex2 ) {
-				ex2.printStackTrace();
-			}
+			FlatLightLaf.install();
 		}
 
 		// create frame